Move property information view screen to own folder
