+++ frontend/src/features/mapSideBar/tabs/propertyDetails/detail/PropertyDetailsTabView.helpers.tsx
