+++ frontend/src/features/mapSideBar/tabs/propertyDetails/detail/PropertyDetailsTabView.test.tsx
