+++ frontend/src/features/mapSideBar/tabs/propertyDetails/detail/PropertyDetailsTabView.tsx
@@ -10,9 +10,9 @@ import React from 'react';
 import { Col, Row } from 'react-bootstrap';
 import { stringToBoolean } from 'utils/formUtils';
 
-import { Section } from '../Section';
-import { SectionField, StyledFieldLabel } from '../SectionField';
-import { InlineContainer, LeftBorderCol, StyledReadOnlyForm } from '../SectionStyles';
+import { Section } from '../../Section';
+import { SectionField, StyledFieldLabel } from '../../SectionField';
+import { InlineContainer, LeftBorderCol, StyledReadOnlyForm } from '../../SectionStyles';
 import { LandMeasurementTable } from './components/LandMeasurementTable';
 import { VolumetricMeasurementTable } from './components/VolumetricMeasurementTable';
 import {
